pyproject.toml
[project]
name = "design-system-agent"
version = "0.1.0"
description = "AI Agent for Design System (Figma + Storybook)"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.0.0",
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gemini = ["google-genai>=1.0.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
